fix warnings

+++ cocos/2d/CCFontFNT.cpp
@@ -293,7 +293,7 @@ std::set<unsigned int>* BMFontConfiguration::parseConfigFile(const std::string& 
     
     auto next = strchr(contents, '\n');
     auto base = contents;
-    int lineLength = 0;
+    intptr_t lineLength = 0;
     int parseCount = 0;
     while (next)
     {
